Electron: Handle drag and dropping notebooks to change the parent

+++ ReactNativeClient/lib/components/screens/notes.js
@@ -107,7 +107,7 @@ class NotesScreenComponent extends BaseScreenComponent {
 	}
 
 	deleteFolder_onPress(folderId) {
-		dialogs.confirm(this, _('Delete notebook? All notes within this notebook will also be deleted.')).then((ok) => {
+		dialogs.confirm(this, _('Delete notebook? All notes and sub-notebooks within this notebook will also be deleted.')).then((ok) => {
 			if (!ok) return;
 
 			Folder.delete(folderId).then(() => {
